you're hot and sassy

+++ springweb2/src/main/java/Springweb/repository/thanhvienRepository.java
@@ -4,6 +4,6 @@ import Springweb.entity.thanhvien;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-public interface thanhvienRepository extends CrudRepository<thanhvien, Long> {
+public interface thanhvienRepository extends CrudRepository<thanhvien, Integer> {
 
 }
