+++ springweb2/src/main/java/Springweb/repository/thietbiRepository.java
@@ -1,9 +1,11 @@
 package Springweb.repository;
 
 import Springweb.entity.thietbi;
+import java.util.Optional;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-public interface thietbiRepository extends CrudRepository<thietbi, Long> {
+public interface thietbiRepository extends CrudRepository<thietbi, Integer> {
 
+    
 }
