+++ springweb2/src/main/java/Springweb/repository/thongtinsdRepository.java
@@ -4,6 +4,6 @@ import Springweb.entity.thongtinsd;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-public interface thongtinsdRepository extends CrudRepository<thongtinsd, Long> {
+public interface thongtinsdRepository extends CrudRepository<thongtinsd, Integer> {
 
 }
