+++ springweb2/src/main/java/Springweb/repository/xulyRepository.java
@@ -4,6 +4,6 @@ import Springweb.entity.xuly;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-public interface xulyRepository extends CrudRepository<xuly, Long>{
+public interface xulyRepository extends CrudRepository<xuly, Integer>{
 
 }
